==> 04_match_sword_to_planet.ipynb <==
import geopandas as gpd
from glob import glob
from tqdm import tqdm
import pandas as pd
import numpy as np
import os

sword_dir = "/nas/cee-water/cjgleason/data/SWORD/SWORDv16/shp"
subdirs = ["AF", "AS", "EU", "NA", "OC", "SA"]

all_sword_fps = []
for subdir in subdirs:
    for fp in glob(os.path.join(sword_dir, subdir, "*reaches*.shp")):
        all_sword_fps.append(fp)
        print(fp)
print(len(all_sword_fps))

# Create mapping of reach_id to sword filepath
all_sword_mapping = None
for fp in tqdm(all_sword_fps):
    data = gpd.read_file(fp)
    mapping_data = data[['reach_id', 'reach_len','width','width_var']]
    mapping_data["sword_fp"] = fp
    if all_sword_mapping is None:
        all_sword_mapping = mapping_data
    else:
        all_sword_mapping = pd.concat([all_sword_mapping, mapping_data])
all_sword_mapping.to_csv("all_sword_mapping_v16.csv", index=False)  


all_sword_mapping[all_sword_mapping["reach_id"]==81210100041]

imgs_dir = "/project/pi_cjgleason_umass_edu/swot_planet/unit/"
ref_df_fp = "/project/pi_cjgleason_umass_edu/swot_planet/yukon_imgs.csv"
ref_df = pd.read_csv(ref_df_fp)

all_planet_reach_ids = np.unique(ref_df["reach_id"].values)
all_planet_reach_ids = [int(x) for x in all_planet_reach_ids]
ctr = 0
for id in all_planet_reach_ids:
    filtered = all_sword_mapping[all_sword_mapping["reach_id"]==id]
    if len(filtered)==0:
        print(id)
        ctr += 1
print(ctr)
print(len(all_planet_reach_ids))

mapping = pd.read_csv("all_sword_mapping_v16.csv")
